Clean up duplicated code
